libstagefright: Fix timestamp corruption due to overflow of 32bit variables
-Timestamp gets corrupted when the calculated value goes beyond
 32 bit. This causes an error during playback and seeking of long
 duration clips.
-Add code to fix timestamp corruption due to overflow of 32bit
 variables.

CRs-Fixed: 265425

+++ media/libstagefright/MPEG4Extractor.cpp
@@ -332,7 +332,7 @@ sp<MetaData> MPEG4Extractor::getTrackMetaData(
         CHECK(track->meta->findCString(kKeyMIMEType, &mime));
         if (!strncasecmp("video/", mime, 6)) {
             uint32_t sampleIndex;
-            uint32_t sampleTime;
+            uint64_t sampleTime;
             if (track->sampleTable->findThumbnailSample(&sampleIndex) == OK
                     && track->sampleTable->getMetaDataForSample(
                         sampleIndex, NULL /* offset */, NULL /* size */,
@@ -1597,7 +1597,7 @@ status_t MPEG4Source::read(
             return err;
         }
 
-        uint32_t sampleTime;
+        uint64_t sampleTime;
         CHECK_EQ((status_t)OK, mSampleTable->getMetaDataForSample(
                     sampleIndex, NULL, NULL, &sampleTime));
 
@@ -1606,7 +1606,7 @@ status_t MPEG4Source::read(
         }
 
 #if 0
-        uint32_t syncSampleTime;
+        uint64_t syncSampleTime;
         CHECK_EQ(OK, mSampleTable->getMetaDataForSample(
                     syncSampleIndex, NULL, NULL, &syncSampleTime));
 
@@ -1628,7 +1628,7 @@ status_t MPEG4Source::read(
 
     off64_t offset;
     size_t size;
-    uint32_t dts;
+    uint64_t dts;
     bool isSyncSample;
     bool newBuffer = false;
     if (mBuffer == NULL) {
